Membangun_model/requirements.txt
mlflow
pandas
numpy
scikit-learn
torch
torchvision
pillow
matplotlib
seaborn
dagshub
python-dotenv